re-fixed logging of sql, set logging to debug
re-fixed logging of sql so now you can see that swear checker doesn't work with the new exchange format of camel messaging. Author (username) returns null when isUserBlocked(string, string) is called. Looks like the prepared statement sql needs to be rewritten, the if statement only succeeds in checking for users who are not blocked.

Signed-off-by: Matthew Stevens <matt@graffitihaus.com>